Add signer skeletons; re-arrange some stuff; add test model
