+++ codegen/smithy-ruby-codegen/build.gradle.kts
@@ -37,3 +37,7 @@ dependencies {
     implementation("software.amazon.smithy:smithy-waiters:${rootProject.extra["smithyVersion"]}")
     implementation("software.amazon.smithy:smithy-protocol-test-traits:${rootProject.extra["smithyVersion"]}")
 }
+
+java.sourceSets["main"].java {
+    srcDirs("model", "src/main/smithy")
+}
