+++ hearth/lib/hearth.rb
@@ -16,6 +16,8 @@ require_relative 'hearth/request'
 require_relative 'hearth/response'
 
 require_relative 'hearth/http'
+require_relative 'hearth/identity'
+require_relative 'hearth/identity_resolver'
 require_relative 'hearth/interceptor'
 require_relative 'hearth/interceptor_context'
 require_relative 'hearth/interceptor_list'
@@ -27,6 +29,8 @@ require_relative 'hearth/number_helper'
 require_relative 'hearth/output'
 require_relative 'hearth/query/param'
 require_relative 'hearth/query/param_list'
+require_relative 'hearth/refreshing_identity'
+require_relative 'hearth/refreshing_identity_resolver'
 require_relative 'hearth/retry'
 require_relative 'hearth/structure'
 require_relative 'hearth/stubbing/client_stubs'
