--- hearth/lib/hearth/auth/http.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'http/api_key_identity'
-require_relative 'http/bearer_token_identity'
-require_relative 'http/login_identity'
