+++ hearth/lib/hearth/auth/scheme.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Hearth
+  module Auth
+    class Scheme
+      def initialize(scheme_id:, identity_resolver:, signer:)
+
+      end
+    end
+  end
+end
