+++ hearth/lib/hearth/identity/anonymous_identity.rb
