+++ hearth/lib/hearth/identity/http_api_key_identity.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module Hearth
+  module Auth
+    # Identity class for API Key authentication.
+    class HTTPApiKeyIdentity < Identity
+      def initialize(key:, **kwargs)
+        super(**kwargs)
+        @key = key
+      end
+
+      # @return [String]
+      attr_reader :key
+    end
+  end
+end
