+++ hearth/lib/hearth/identity/http_bearer_token_identity.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module Hearth
+  module Auth
+    # Identity class for bearer token authentication.
+    class HTTPBearerTokenIdentity < Identity
+      def initialize(token:, **kwargs)
+        super(**kwargs)
+        @token = token
+      end
+
+      # @return [String]
+      attr_reader :token
+    end
+  end
+end
