+++ hearth/lib/hearth/identity/http_login_identity.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+module Hearth
+  module Auth
+    # Identity class for login authentication.
+    class HTTPLoginIdentity < Identity
+      def initialize(username:, password:, **kwargs)
+        super(**kwargs)
+        @username = username
+        @password = password
+      end
+
+      # @return [String]
+      attr_reader :username
+
+      # @return [String]
+      attr_reader :password
+    end
+  end
+end
