+++ hearth/lib/hearth/signers/http_basic.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module Hearth
+  module Signers
+    # A signer that signs requests using the HTTP Basic Auth scheme.
+    class HTTPBasic
+      def sign(request:, identity:, properties: {})
+        # TODO
+      end
+    end
+  end
+end
