+++ hearth/lib/hearth/signers/http_digest.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module Hearth
+  module Signers
+    # A signer that signs requests using the HTTP Digest Auth scheme.
+    class HTTPDigest
+      def sign(request:, identity:, properties: {})
+        # TODO
+      end
+    end
+  end
+end
